add loading animation to application and refactored files
