+++ src/FlightCard/FlightCard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const FlightCard = () => {
+  return (
+    <div>FlightCard</div>
+  )
+}
+
+export default FlightCard
