+++ src/FlightsContainer/FlightsContainer.css
@@ -1,0 +1,6 @@
+.widget {
+    background-color: black;
+    opacity: 85%;
+    height: 29rem;
+    width: 50rem;
+  }
