+++ src/Loader/Loader.js
@@ -1,0 +1,10 @@
+import React from "react";
+import "./Loader.css"
+
+const Loader = () => {
+    return (
+        <div className="loader">Loading...</div>
+    )
+}
+
+export default Loader
